# apps/auth-service/docs/openapi/openapi-docs.json
{"openapi":"3.0.0","paths":{"/auth/username/available":{"get":{"operationId":"AuthController_verifyIfUsernameIsAvailable","summary":"Retrieves information if the provided username is already registered","parameters":[{"name":"username","required":true,"in":"query","description":"Username to be verified","example":"anderson","schema":{"type":"string"}},{"name":"X-Api-Version","in":"header","description":"Sets the API version","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Information about the username availability","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyUsernameSerializer"}}}},"400":{"description":"Error validating request input data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyUsernameError400Serializer"}}}},"500":{"description":"The server has encountered a situation it does not know how to handle. See server logs for details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultError500Serializer"}}}},"502":{"description":"Internal data processing error. Probably a database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultError502Serializer"}}}}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"Uses username and password to get a JWT token to use in future API calls","parameters":[{"name":"X-Api-Version","in":"header","description":"Sets the API version","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginBodyDto"}}}},"responses":{"200":{"description":"JWT Access Token to use in future API calls","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginSerializer"}}}},"400":{"description":"Error validating request input data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginError400Serializer"}}}},"401":{"description":"User do not exists, exists and is inactive or the password is wrong","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginError401Serializer"}}}},"500":{"description":"The server has encountered a situation it does not know how to handle. See server logs for details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultError500Serializer"}}}},"502":{"description":"Internal data processing error. Probably a database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultError502Serializer"}}}}},"tags":["auth"]}},"/auth/signup":{"post":{"operationId":"AuthController_signup","summary":"Creates a new user","parameters":[{"name":"X-Api-Version","in":"header","description":"Sets the API version","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpBodyDto"}}}},"responses":{"200":{"description":"ID for the created user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpSerializer"}}}},"400":{"description":"Error validating request input data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpError400Serializer"}}}},"409":{"description":"User cannot be created. It already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpError409Serializer"}}}},"500":{"description":"The server has encountered a situation it does not know how to handle. See server logs for details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultError500Serializer"}}}},"502":{"description":"Internal data processing error. Probably a database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultError502Serializer"}}}}},"tags":["auth"]}}},"info":{"title":"Auth Service","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"VerifyUsernameSerializer":{"type":"object","properties":{"available":{"type":"boolean","description":"Informs if the username is available for use","example":true}},"required":["available"]},"VerifyUsernameError400Serializer":{"type":"object","properties":{"statusCode":{"type":"number","description":"HTTP code","example":400},"message":{"description":"","example":["username should not be empty"],"type":"array","items":{"type":"string"}},"error":{"type":"string","description":"HTTP error message","example":"Bad Request"}},"required":["statusCode","message","error"]},"DefaultError500Serializer":{"type":"object","properties":{"statusCode":{"type":"number","description":"HTTP code","example":500},"message":{"type":"string","description":"Additional error message","example":"Cannot set properties of undefined"},"error":{"type":"string","description":"HTTP error message","example":"Internal Server Error"}},"required":["statusCode","message","error"]},"DefaultError502Serializer":{"type":"object","properties":{"statusCode":{"type":"number","description":"HTTP code","example":502},"message":{"type":"string","description":"HTTP error message","example":"Bad Gateway"}},"required":["statusCode","message"]},"LoginBodyDto":{"type":"object","properties":{"username":{"type":"string","description":"Username","example":"anderson"},"password":{"type":"string","description":"User password","example":"te$st@12%34"}},"required":["username","password"]},"LoginSerializer":{"type":"object","properties":{"accessToken":{"type":"string","description":"Token to be used at the private endpoints","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFuZGVyc29uIiwic3ViIjoxLCJpYXQiOjE2ODM4MzAyNTEsImV4cCI6MTY4MzgzMDMxMX0.eN5Cv2tJ0HGlVNKMtPv5VPeCIA7dd4OEA-8Heh7OJ_c"}},"required":["accessToken"]},"LoginError400Serializer":{"type":"object","properties":{"statusCode":{"type":"number","description":"HTTP code","example":400},"message":{"description":"","example":["username should not be empty"],"type":"array","items":{"type":"string"}},"error":{"type":"string","description":"HTTP error message","example":"Bad Request"}},"required":["statusCode","message","error"]},"LoginError401Serializer":{"type":"object","properties":{"statusCode":{"type":"number","description":"HTTP code","example":400},"error":{"type":"string","description":"HTTP error message","example":"Unauthorized"}},"required":["statusCode","error"]},"SignUpBodyDto":{"type":"object","properties":{"userId":{"type":"string","description":"User id from the main database","example":"76c9d285-ab38-48e2-b97b-7556150c11ce"},"username":{"type":"string","description":"User name","example":"anderson"},"password":{"type":"string","description":"Password in plain text","example":"test@1234"}},"required":["userId","username","password"]},"SignUpSerializer":{"type":"object","properties":{}},"SignUpError400Serializer":{"type":"object","properties":{"statusCode":{"type":"number","description":"HTTP code","example":400},"message":{"description":"","example":["username should not be empty"],"type":"array","items":{"type":"string"}},"error":{"type":"string","description":"HTTP error message","example":"Bad Request"}},"required":["statusCode","message","error"]},"SignUpError409Serializer":{"type":"object","properties":{"statusCode":{"type":"number","description":"HTTP code","example":409},"message":{"type":"string","description":"HTTP error message","example":"Conflict"}},"required":["statusCode","message"]}}}}